+++ crud.go
@@ -44,6 +44,7 @@ func NewCrud(params mctypes.CrudParamsType, options mctypes.CrudOptionsType) (cr
 	crudInstance.AccessTable = options.AccessTable
 	crudInstance.RoleTable = options.RoleTable
 	crudInstance.UserTable = options.UserTable
+	crudInstance.UserProfileTable = options.UserProfileTable
 	crudInstance.ServiceTable = options.ServiceTable
 	crudInstance.AuditDb = options.AuditDb
 	crudInstance.AccessDb = options.AccessDb
@@ -53,7 +54,7 @@ func NewCrud(params mctypes.CrudParamsType, options mctypes.CrudOptionsType) (cr
 	crudInstance.LogUpdate = options.LogUpdate
 	crudInstance.LogDelete = options.LogDelete
 	crudInstance.CheckAccess = options.CheckAccess // Dec 09/2020: user to implement auth as a middleware
-	crudInstance.CacheExpire = options.CacheExpire	// cache expire in secs
+	crudInstance.CacheExpire = options.CacheExpire // cache expire in secs
 	// Compute HashKey from TableName, QueryParams, SortParams, ProjectParams and RecordIds
 	qParam, _ := json.Marshal(params.QueryParams)
 	sParam, _ := json.Marshal(params.SortParams)
@@ -73,6 +74,9 @@ func NewCrud(params mctypes.CrudParamsType, options mctypes.CrudOptionsType) (cr
 	}
 	if crudInstance.UserTable == "" {
 		crudInstance.UserTable = "users"
+	}
+	if crudInstance.UserProfileTable == "" {
+		crudInstance.UserProfileTable = "user_profile"
 	}
 	if crudInstance.ServiceTable == "" {
 		crudInstance.ServiceTable = "services"
@@ -96,7 +100,7 @@ func NewCrud(params mctypes.CrudParamsType, options mctypes.CrudOptionsType) (cr
 	}
 
 	if crudInstance.CacheExpire <= 0 {
-		crudInstance.CacheExpire = 300		// 300 secs, 5 minutes
+		crudInstance.CacheExpire = 300 // 300 secs, 5 minutes
 	}
 
 	// Audit/TransLog instance